Separate contract subpackages as the contract package is already large and will need to accomodate many more different service types
