+++ core/contract/src/main/java/org/eclipse/dataspaceconnector/contract/ContractServiceExtension.java
@@ -14,10 +14,13 @@
 
 package org.eclipse.dataspaceconnector.contract;
 
+import org.eclipse.dataspaceconnector.contract.agent.ParticipantAgentServiceImpl;
+import org.eclipse.dataspaceconnector.contract.offer.ContractOfferServiceImpl;
+import org.eclipse.dataspaceconnector.contract.offer.NullContractOfferFramework;
 import org.eclipse.dataspaceconnector.spi.asset.AssetIndex;
-import org.eclipse.dataspaceconnector.spi.contract.ContractOfferFramework;
-import org.eclipse.dataspaceconnector.spi.contract.ContractOfferService;
-import org.eclipse.dataspaceconnector.spi.contract.ParticipantAgentService;
+import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferFramework;
+import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferService;
+import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgentService;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
