+++ core/contract/src/main/java/org/eclipse/dataspaceconnector/contract/NullAssetIndex.java
@@ -1,3 +1,16 @@
+/*
+ *  Copyright (c) 2021 Daimler TSS GmbH
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Daimler TSS GmbH - Initial API and Implementation
+ *
+ */
 package org.eclipse.dataspaceconnector.contract;
 
 import org.eclipse.dataspaceconnector.spi.asset.AssetIndex;
@@ -7,7 +20,7 @@ import org.eclipse.dataspaceconnector.spi.types.domain.asset.Asset;
 import java.util.stream.Stream;
 
 /**
- * NullObject of the {@link AssetIndex}
+ * A stub asset index that contains no assets.
  */
 public class NullAssetIndex implements AssetIndex {
