+++ core/contract/src/main/java/org/eclipse/dataspaceconnector/contract/agent/ParticipantAgentServiceImpl.java
@@ -11,11 +11,11 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.contract;
+package org.eclipse.dataspaceconnector.contract.agent;
 
-import org.eclipse.dataspaceconnector.spi.contract.ParticipantAgent;
-import org.eclipse.dataspaceconnector.spi.contract.ParticipantAgentService;
-import org.eclipse.dataspaceconnector.spi.contract.ParticipantAgentServiceExtension;
+import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgentService;
+import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgentServiceExtension;
 import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
 
 import java.util.ArrayList;
