+++ data-protocols/ids/ids-spi/src/main/java/org/eclipse/dataspaceconnector/ids/spi/types/container/OfferedAsset.java
@@ -1,7 +1,7 @@
 package org.eclipse.dataspaceconnector.ids.spi.types.container;
 
 import org.eclipse.dataspaceconnector.spi.types.domain.asset.Asset;
-import org.eclipse.dataspaceconnector.spi.types.domain.contract.ContractOffer;
+import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractOffer;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
