+++ samples/demo-contract-framework/src/main/java/org/eclipse/dataspaceconnector/demo/contract/offer/ContractOfferController.java
@@ -1,4 +1,4 @@
-package org.eclipse.dataspaceconnector.demo.contracts;
+package org.eclipse.dataspaceconnector.demo.contract.offer;
 
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.GET;
@@ -6,14 +6,15 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import org.eclipse.dataspaceconnector.spi.contract.ContractOfferQuery;
-import org.eclipse.dataspaceconnector.spi.contract.ContractOfferService;
+import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferQuery;
+import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferService;
 
 import java.util.stream.Collectors;
 
 @Consumes({MediaType.APPLICATION_JSON})
 @Produces({MediaType.APPLICATION_JSON})
 @Path("/offers")
+@Deprecated
 public class ContractOfferController {
 
     private final ContractOfferService contractOfferService;
