+++ samples/demo-contract-framework/src/main/java/org/eclipse/dataspaceconnector/demo/contract/offer/PublicContractOfferFramework.java
@@ -11,15 +11,15 @@
  *       Daimler TSS GmbH - Initial API and Implementation
  *       Microsoft Corporation - Refactoring
  */
-package org.eclipse.dataspaceconnector.demo.contracts;
+package org.eclipse.dataspaceconnector.demo.contract.offer;
 
 import org.eclipse.dataspaceconnector.policy.model.Action;
 import org.eclipse.dataspaceconnector.policy.model.Permission;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.policy.model.PolicyType;
-import org.eclipse.dataspaceconnector.spi.contract.ContractDefinition;
-import org.eclipse.dataspaceconnector.spi.contract.ContractOfferFramework;
-import org.eclipse.dataspaceconnector.spi.contract.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.contract.offer.ContractDefinition;
+import org.eclipse.dataspaceconnector.spi.contract.offer.ContractOfferFramework;
+import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgent;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
