+++ samples/demo-contract-framework/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -12,5 +12,5 @@
 #
 #
 
-org.eclipse.dataspaceconnector.demo.contracts.DemoContractOfferFrameworkExtension
+org.eclipse.dataspaceconnector.demo.contract.offer.DemoContractOfferFrameworkExtension
