+++ samples/demo-contract-framework/src/test/java/org/eclipse/dataspaceconnector/demo/contracts/DemoContractOfferFrameworkExtensionTest.java
@@ -22,7 +22,7 @@ import okhttp3.RequestBody;
 import okhttp3.Response;
 import org.easymock.EasyMock;
 import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
-import org.eclipse.dataspaceconnector.spi.contract.ParticipantAgentService;
+import org.eclipse.dataspaceconnector.spi.contract.agent.ParticipantAgentService;
 import org.eclipse.dataspaceconnector.spi.system.ConfigurationExtension;
 import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 import org.eclipse.dataspaceconnector.spi.types.domain.asset.Asset;
