+++ spi/src/main/java/org/eclipse/dataspaceconnector/spi/contract/agent/ParticipantAgentService.java
@@ -11,7 +11,7 @@
  *       Microsoft Corporation - initial API and implementation
  *
  */
-package org.eclipse.dataspaceconnector.spi.contract;
+package org.eclipse.dataspaceconnector.spi.contract.agent;
 
 import org.eclipse.dataspaceconnector.spi.iam.ClaimToken;
